+++ .env
@@ -1,0 +1,6 @@
+EXPO_PUBLIC_APIKEY=AIzaSyCyg3VcXEpPugt9HutUDyTfe5OBnJ5PLB8
+EXPO_PUBLIC_AUTHDOMAIN=myexchange-b2d22.firebaseapp.com
+EXPO_PUBLIC_PROJECTID=myexchange-b2d22
+EXPO_PUBLIC_STORAGEBUCKET=myexchange-b2d22.appspot.com
+EXPO_PUBLIC_MESSAGINGSENDERID=67656550212
+EXPO_PUBLIC_APPID=1:67656550212:web:a278bb9bf51bf397b87911
