+++ App.js
@@ -1,6 +1,6 @@
 import { StatusBar } from 'expo-status-bar';
 import { StyleSheet, Text, View } from 'react-native';
-import AppNavigation from './navigation';
+import AppNavigation from './src/routes/stackRoutes';
 
 export default function App() {
   return (
