+++ src/screens/HomePage.js
@@ -1,4 +1,4 @@
-import React, {useState, useEffect} from 'react';
+import React, { useState, useEffect } from 'react';
 import { View, Text, Image, TouchableOpacity, Linking } from 'react-native';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import { SafeAreaView } from 'react-native-safe-area-context';
@@ -10,28 +10,9 @@ import Icon from 'react-native-vector-icons/FontAwesome';
 
 export default function HomePage() {
     const navigation = useNavigation();
-    /* const [name, setName] = useState('');
 
-    useEffect(() => {
-        const fetchName = async () => {
-            const user = auth.currentUser;
-            if (user) {
-                const docRef = doc(db, "users", user.uid);
-                const docSnap = await getDoc(docRef);
-
-                if (docSnap.exists()) {
-                    setName(docSnap.data().name);
-                } else {
-                    console.log("No such document!");
-                }
-            }
-        };
-
-        fetchName();
-    }, []); */
-
-    const handleLogout = async ()=>{
-      await signOut(auth);
+    const handleLogout = async () => {
+        await signOut(auth);
     }
     const user = {
         name: "Vagner Pires",
@@ -43,32 +24,32 @@ export default function HomePage() {
         originCity: 'Novo Hamburgo',
         originCountry: 'Brazil',
     };
-    const handleEditProfile = async ()=>{
-        await navigation.navigate('EditProfile');
+    const handleEditProfile = () => {
+        navigation.navigate('EditProfile');
     };
 
-    const handleNavigateToSchool = async ()=>{
-        await navigation.navigate('School');
+    const handleNavigateToSchool = () => {
+        navigation.navigate('School');
     };
 
-    const handleNavigateToAirline = async ()=>{
-        await navigation.navigate('Airline');
+    const handleNavigateToAirline = () => {
+        navigation.navigate('Airline');
     };
 
     const openWhatsAppChat = () => {
         let phoneNumber = '+353 89 975 3246';
         let message = 'Hello, I would like to chat with you on WhatsApp!';
         let url = `whatsapp://send?text=${encodeURIComponent(message)}&phone=${phoneNumber}`;
-    
+
         Linking.canOpenURL(url)
-        .then((supported) => {
-            if (!supported) {
-                console.log("Can't handle url: " + url);
-            } else {
-                return Linking.openURL(url);
-            }
-        })
-        .catch((err) => console.error('An error occurred', err));
+            .then((supported) => {
+                if (!supported) {
+                    console.log("Can't handle url: " + url);
+                } else {
+                    return Linking.openURL(url);
+                }
+            })
+            .catch((err) => console.error('An error occurred', err));
     }
 
     const [liked, setLiked] = useState(false);
@@ -83,55 +64,55 @@ export default function HomePage() {
         }
         loadFavorites();
     }, []);
-    
+
     const toggleLike = async () => {
         setLiked(!liked);
-            
+
         const updatedFavorites = liked
             ? favorites.filter((favorite) => favorite !== user.id)
             : [...favorites, user];
         setFavorites(updatedFavorites);
         await AsyncStorage.setItem('favorites', JSON.stringify(updatedFavorites));
     };
-    
+
     return (
-      <SafeAreaView style={styles.container}>
-        <TouchableOpacity onPress={handleLogout} style={styles.logout}>
-            <Text style={styles.editButtonText}>LOGOUT</Text>
-        </TouchableOpacity>
-        <TouchableOpacity onPress={handleEditProfile} style={styles.editButton}>
-            <Text style={styles.editButtonText}>MY PROFILE</Text>
-        </TouchableOpacity>
-        <Image source={require('../assets/profile.png')} style={styles.userPhoto} />
-        <Text style={[styles.userName, styles.userAge]}>{user.name} - {user.age}</Text>
-        <View style={styles.separator}></View>
-        <Image source={require('../assets/flag.png')} style={styles.userCountry} />
-        <Text style={styles.userCity}>{user.city}</Text>
-        <Text style={styles.userArrivalDate}>Arrival Date: {user.arrivalDate}</Text>
-        <View style={styles.schoolContainer}>
-            <Text style={styles.userSchool}>School: </Text>
-            <TouchableOpacity onPress={handleNavigateToSchool}>
-            <Text style={styles.linkText}>{user.school}</Text>
+        <SafeAreaView style={styles.container}>
+            <TouchableOpacity onPress={handleLogout} style={styles.logout}>
+                <Text style={styles.editButtonText}>LOGOUT</Text>
             </TouchableOpacity>
-        </View>
-        <View style={styles.airlineContainer}>
-            <Text style={styles.userAirlines}>Airline: </Text>
-            <TouchableOpacity onPress={handleNavigateToAirline}>
-            <Text style={styles.linkText}>{user.airlines}</Text>
+            <TouchableOpacity onPress={handleEditProfile} style={styles.editButton}>
+                <Text style={styles.editButtonText}>MY PROFILE</Text>
             </TouchableOpacity>
-        </View>
-        <View style={styles.separator2}></View>
-        <Text style={styles.userOriginCity}>{user.originCity}</Text>
-        <Text style={styles.userOriginCountry}>{user.originCountry}</Text>
-        <TouchableOpacity onPress={openWhatsAppChat}>
-            <Image source={require('../assets/whatsapp.png')} style={styles.whatsappLogo} />
-        </TouchableOpacity>
-        <Footer navigation={navigation} />
-      </SafeAreaView>
+            <Image source={require('../../assets/profile.png')} style={styles.userPhoto} />
+            <Text style={[styles.userName, styles.userAge]}>{user.name} - {user.age}</Text>
+            <View style={styles.separator}></View>
+            <Image source={require('../../assets/flag.png')} style={styles.userCountry} />
+            <Text style={styles.userCity}>{user.city}</Text>
+            <Text style={styles.userArrivalDate}>Arrival Date: {user.arrivalDate}</Text>
+            <View style={styles.schoolContainer}>
+                <Text style={styles.userSchool}>School: </Text>
+                <TouchableOpacity onPress={handleNavigateToSchool}>
+                    <Text style={styles.linkText}>{user.school}</Text>
+                </TouchableOpacity>
+            </View>
+            <View style={styles.airlineContainer}>
+                <Text style={styles.userAirlines}>Airline: </Text>
+                <TouchableOpacity onPress={handleNavigateToAirline}>
+                    <Text style={styles.linkText}>{user.airlines}</Text>
+                </TouchableOpacity>
+            </View>
+            <View style={styles.separator2}></View>
+            <Text style={styles.userOriginCity}>{user.originCity}</Text>
+            <Text style={styles.userOriginCountry}>{user.originCountry}</Text>
+            <TouchableOpacity onPress={openWhatsAppChat}>
+                <Image source={require('../../assets/whatsapp.png')} style={styles.whatsappLogo} />
+            </TouchableOpacity>
+            <Footer navigation={navigation} />
+        </SafeAreaView>
     )
 }
 
-const Footer = ({navigation}) => {
+const Footer = ({ navigation }) => {
     const handleHomePage = () => {
         navigation.navigate('Home');
     };
@@ -217,7 +198,7 @@ const styles = {
         position: 'absolute',
         marginTop: 470,
         right: 30,
-      },
+    },
     separator: {
         position: 'absolute',
         width: 335,
